backend/admin-auth/requirements.txt
pyjwt>=2.8.0
bcrypt>=4.1.0
